requirements.txt
django>=4.0.3,<4.1
djangorestframework>=3.14.0,<3.15
psycopg2>=2.8.6,<2.9
drf-spectacular
Pillow